fix getconfig
