+++ css/package.json
@@ -7,7 +7,7 @@
     "src"
   ],
   "scripts": {
-    "build": "tsc"
+    "build": "tsc -p ./tsconfig.build.json"
   },
   "dependencies": {
     "@windicss/plugin-interaction-variants": "^1.0.0",
