+++ css/src/windi.config.ts
@@ -1,9 +1,14 @@
 import { defineConfig } from 'windicss/helpers'
 // @ts-ignore
-import InteractionVariants from '@windicss/plugin-interaction-variants'
+import PluginInteractionVariants from '@windicss/plugin-interaction-variants'
+import PluginFilters from 'windicss/plugin/filters'
 import { IconDuotoneColorsPlugin, IconExtractor } from './icon-color-plugins'
 import { colors } from './colors'
 import { shortcuts } from './shortcuts'
+
+// give the plugins a name for debugging purposes
+PluginInteractionVariants.config = { name: 'windi-interaction-variants' }
+PluginFilters.config = { name: 'windi-filters' }
 
 export default defineConfig({
   // This adds !important to all utility classes.
@@ -44,11 +49,7 @@ export default defineConfig({
       'no-hover',
     ],
   },
-  plugins: [
-    IconDuotoneColorsPlugin,
-    InteractionVariants,
-    require('windicss/plugin/filters'),
-  ],
+  plugins: [IconDuotoneColorsPlugin, PluginInteractionVariants, PluginFilters],
   shortcuts,
   extract: {
     extractors: [IconExtractor],
